Make the code work with python3. Requires the future module.
Most of the work was just turning floats into ints for numpy, and updating the
syntax for exceptions. I had to remove "ionFR_2018" in a path, which I assume is
mistakingly left in.

I tested the output of:
./ionFRM.py 08h37m05.6s+06d10m14.5s 52d54m54.6sn 6d52m11.7se 2011-10-20T00:00:00 test/codg2930.11i

with python 2 and 3, and the outputs are the same, except for slight precision
differences in the floats.